+++ tests/testthat/test_hits.R
@@ -19,3 +19,50 @@ test_that(
     expect_equal(y@stat[id == "242"][["count"]], count(partition("REUTERS", id = "242"), query = "oil")[["count"]])
   }
 )
+
+test_that(
+  "hits for nested scenario",
+  {
+    testthat::skip_on_cran()
+    use("GermaParl2")
+    
+    # we want to ensure that the order is independent from the order of 
+    # subsetting operations
+    
+    hits1 <- corpus("GERMAPARL2MINI") |>
+      subset(p_type == "speech") %>% 
+      subset(speaker_party %in% c("CDU", "CSU", "SPD")) |>
+      split(s_attribute = "speaker_party", verbose = FALSE) %>% 
+      hits(query = tm::stopwords("de")[1:10], cqp = FALSE, verbose = FALSE) %>% 
+      as.data.table()
+    
+    hits2 <- corpus("GERMAPARL2MINI") |>
+      subset(speaker_party %in% c("CDU", "CSU", "SPD")) |>
+      split(s_attribute = "speaker_party") %>% 
+      subset(p_type == "speech", verbose = FALSE) %>% 
+      hits(query = tm::stopwords("de")[1:10], cqp = FALSE, verbose = FALSE) %>% 
+      as.data.table()
+    
+    testthat::expect_identical(hits1, hits2)
+    
+    hits3 <- corpus("GERMAPARL2MINI") |>
+      subset(speaker_party == "CDU") |>
+      subset(p_type == "speech", verbose = FALSE) %>% 
+      hits(query = tm::stopwords("de")[1:10], cqp = FALSE, verbose = FALSE) %>% 
+      as.data.table()
+    
+    hits4 <- corpus("GERMAPARL2MINI") |>
+      subset(p_type == "speech") %>% 
+      subset(speaker_party == "CDU") |>
+      hits(query = tm::stopwords("de")[1:10], cqp = FALSE, verbose = FALSE) %>% 
+      as.data.table()
+    
+    testthat::expect_identical(hits3, hits4)
+    
+    testthat::expect_identical(
+      hits2[partition == "CDU"][, c("query", "count")],
+      hits3[count > 0][,c("query", "count")]
+    )
+    
+  }
+)